backend/app/requirements.txt
fastapi
uvicorn
motor
httpx
python-dotenv
redis
python-jose
rapidfuzz